requirements.txt
flask
flask_restful
opencv-python
pytesseract
Pillow
firebase-admin
google-cloud